Add the concept of server ports to the firewall.
A server port is a local TCP port which is listened to by an
application.

Any new TCP connection to a local server port (detected through an
incoming SYN) causes the creation of a firewall entry.

There is no equivalent concept for UDP, as we do not currently have any
use-case for a UDP server that cannot be handled with custom rules (NTP,
DNS, DHCP, QUIC).

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@ubc.ca>