Implement the new network stack server API.
The new server API allows application-level code to create listening TCP
sockets bound to given server ports, and accept connections on a
listening socket.

The permission to open a given server port is encoded in bind
capabilities: capabilities sealed with a NetAPI-owned key that permit
the opening of a TCP server port on either IPv4 or IPv6.

Important implementation point here: we use socket callbacks to handle
the situation where a three-way TCP handshake initiated by a peer on a
listening socket fails. In this case the firewall hole will be closed by
the socket disconnection callback and not by a call to
`FreeRTOS_closesocket`. See comment in `network_wrapper.cc` for more
details. This behavior can easily be tested by sending a lone TCP SYN,
e.g., through:

   hping3 -c 1 -S -p $port $ip

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@ubc.ca>