Fix type mismatch in `DECLARE_AND_DEFINE_CONNECTION_CAPABILITY`.
The definition of the struct in `struct ConnectionCapability` and in
`DECLARE_AND_DEFINE_CONNECTION_CAPABILITY` differ in the type of `port`.

The struct definition uses `uint16_t` and the macro `short`.

`short` is signed and may be longer that 16 bits. This is not a problem
right now but may lead to portability issues.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@ubc.ca>